requirements.txt
pandas
google_play_scraper
numpy
matplotlib
seaborn
textBlob
nltk
scikit-learn
gensim
spacy
psycopg2-binary